==> 04-Analysis.Rmd <==
# Analyses

```{r, include = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>",
  cache = FALSE,
  warning = FALSE,
  message = FALSE,
  # Save more figure formats
  dev = c("png", "pdf", "svg", "tiff"),
  fig.ext = c("png", "pdf", "svg", "tiff"),
  fig.retina = 2
)
```

```{r}
library(pacman)
p_load(
  here,
  knitr,
  patchwork,
  scales,
  broom,
  ggstance,
  tidyverse
)
```

```{r}
theme_set(
  theme_classic(base_line_size = .25, base_rect_size = 0) +
    theme(
      strip.text = element_text(size = rel(1)),
      strip.background = element_blank(),
      legend.position = "right"
    )
)
colors <- c("navyblue", "#E60012")
```

```{r}
ac <- read_rds(here("data/noa/ac-excluded.rds"))
pvz <- read_rds(here("data/ea/pvz-excluded.rds"))
ac2 <- read_rds(here("data/noa/ac.rds"))
pvz2 <- read_rds(here("data/ea/pvz.rds"))
```

## Create joint dataset

Create harmonized datasets for easier analysis

```{r}
ac <- ac %>% 
  select(
    player_id,
    spane_balance, autonomy, 
    competence, relatedness, enjoyment, 
    extrinsic, active_play, Hours
  )
pvz <- pvz %>% 
  select(
    player_id, spane_balance, autonomy, 
    competence, relatedness, enjoyment, 
    extrinsic, active_play, Hours
  )
dat <- bind_rows(pvz, ac, .id = "Game") %>% 
  mutate(Game = factor(Game, labels = c("PvZ", "AC:NH")))
```

Game time is in units of 10 hours to make the size of the coefficients bigger and thus easier to interpret e.g. when shown with 2 decimal points.

```{r}
dat$Hours10 <- dat$Hours / 10
dat$active_play10 <- dat$active_play / 10
```

## RQ1: Time and well-being

### Objective vs subjective game time

Describe subjective and objective time difference. These numbers are a bit confusing because the means include everyone, but difference only those who had both values (so cannot compute differences from means)

```{r}
dat %>% 
  group_by(Game) %>%
  mutate(difference = active_play-Hours) %>% 
  summarise(
    across(
      c(Hours, active_play, difference), 
      list(m = ~mean(.x, na.rm = T), s = ~sd(.x, na.rm = T))
      )
    )
```

```{r}
p0 <- dat %>% 
  ggplot(aes(Hours, active_play, col = Game)) +
  scale_color_manual(values = colors) +
  geom_point(shape = 1, alpha = .5, size = .5) +
  scale_y_continuous(breaks = pretty_breaks()) +
  scale_x_continuous(breaks = pretty_breaks()) +
  geom_smooth(method = "lm", col = "black", size = .5, alpha = .25) +
  theme(aspect.ratio = 1) +
  guides(
    color = guide_legend(
      override.aes = list(size = 3, shape = 16, alpha = 1)
      )
    ) + 
  facet_wrap("Game", scales = "free", nrow = 2)
```

```{r}
p1 <- p0 + geom_abline(lty = 2, size = .25)
```

Model fitted separately to both datasets

```{r}
res <- function(model) {
  out1 <- tidy(model, conf.int = TRUE) %>% 
    select(-statistic) %>% 
    rename(SE = std.error)
  out2 <- glance(model) %>% 
    select(1,2, nobs) %>% 
    rename(r2 = r.squared, r2a = adj.r.squared)
  bind_cols(out1, out2)
}
dat %>% 
  group_by(Game) %>% 
  group_modify(~res(lm(active_play ~ Hours, data = .x))) %>% 
  kable(digits = c(0,2,2,1,3,2,2,2,2,1))
```

### Objective time and SWB

```{r}
p2 <- p0 + aes(y = spane_balance)
```

Model fitted separately to both datasets

```{r}
dat %>% 
  group_by(Game) %>% 
  group_modify(~res(lm(scale(spane_balance) ~ Hours10, data = .x))) %>% 
  kable(digits = c(0,2,2,1,3,2,2,2,2,1))
```

A separate figure of just this

```{r objective-time-figure}
foo <- function(game, n) {
  tmp <- dat %>% 
    mutate(Game2 = factor(Game, labels = c("Plants vs. Zombies: Battle for Neighborville", "Animal Crossing: New Horizons"))) %>% 
    filter(Game == game)
  tmp %>% 
    ggplot(aes(Hours, spane_balance)) +
    geom_point(alpha = .5, size = 1.75, color = colors[n]) +
    scale_y_continuous(
      "Player wellbeing",
      breaks = pretty_breaks()
    ) +
    scale_x_continuous(
      str_glue("Hours played\n{unique(tmp$Game2)}"),
      breaks = pretty_breaks()
    ) +
    geom_smooth(
      method = "lm", size = .75, alpha = .2,
      col = colors[n], fill = colors[n]
    ) +
    theme(aspect.ratio = 1, legend.position = "none", strip.text = element_blank()) +
    facet_wrap("Game2", scales = "free", nrow = 1)
}
foo("PvZ", 1) | foo("AC:NH", 2)
```

### Subjective time and SWB

```{r}
p3 <- p0 + aes(x = active_play, y = spane_balance)
```

```{r}
dat %>% 
  group_by(Game) %>% 
  group_modify(~res(lm(scale(spane_balance) ~ active_play10, data = .x))) %>% 
  kable(digits = c(0,2,2,1,3,2,2,2,2,1))
```

### Figure

```{r rq1-figure, fig.height = 4}
(p1 + 
   labs(x = "Hours played", y = "Estimated hours") +
   theme(legend.position = "none") | 
  p2 + 
   labs(x = "Hours played", y = "Well-being") +
   theme(legend.position = "none") | 
  p3 + 
   labs(x = "Estimated hours", y = "Well-being") +
   theme(legend.position = "right")
 ) &
  plot_annotation(tag_levels = "A") &
  theme(strip.text = element_blank())
```


## RQ2: Well-being and motivation

```{r}
# Nice names for plots
dat <- dat %>% rename_all(str_to_title)

dat <- dat %>% 
  rename(Intrinsic = Enjoyment)

# Standardize everything
# Hours (centered) is divided by 10 to put on same scale with others.
dat <- dat %>%
  group_by(Game) %>% 
  mutate(
    across(Spane_balance:Extrinsic, ~as.numeric(scale(., T, T))),
    Hours = as.numeric(scale(Hours10, T, F))
    )

# Fit all models to both datasets
xs <- dat %>% 
  group_by(Game) %>% 
  nest() %>% 
  mutate(
    # Omnibus model where variables work together
    x = map(data, ~lm(Spane_balance ~ (Autonomy + Competence + Relatedness + Intrinsic + Extrinsic) * Hours, data = .x)),
    # Separate models for variables to work alone
    x1 = map(data, ~lm(Spane_balance ~ Autonomy * Hours, data = .x)),
    x2 = map(data, ~lm(Spane_balance ~ Competence * Hours, data = .x)),
    x3 = map(data, ~lm(Spane_balance ~ Relatedness * Hours, data = .x)),
    x4 = map(data, ~lm(Spane_balance ~ Intrinsic * Hours, data = .x)),
    x5 = map(data, ~lm(Spane_balance ~ Extrinsic * Hours, data = .x))
  )
```

```{r rq2-figure, fig.height = 3.8}
tmp <- xs %>% 
  pivot_longer(x:x5, names_to = "Model") %>% 
  mutate(Model = ifelse(Model=="x", 'Omnibus', 'Unique')) %>% 
  mutate(out = map(value, ~tidy(., conf.int = TRUE))) %>% 
  unnest(out) %>% 
  filter(!(term %in% c('(Intercept)'))) %>%
  mutate(term = fct_rev(fct_inorder(term))) %>% 
  mutate(
    Type = factor(
      str_detect(term, ":"), labels = c("Main effect", "Moderation")
    )
  ) 

p1 <- tmp %>% 
  filter(Model == "Omnibus") %>% 
  drop_na() %>% 
  ggplot(aes(estimate, term, col = Game)) +
  scale_color_manual(values = colors) +
  scale_x_continuous(
    'Regression coefficient', breaks = pretty_breaks()
  ) +
  geom_vline(xintercept = 0, lty = 2, size = .2) +
  geom_pointrangeh(
    aes(xmin = conf.low, xmax = conf.high), 
    size = .4, position = position_dodge2v(.4)
    ) +
  theme(
    axis.title.y = element_blank(), 
    strip.text = element_blank()
    )
p1
map(xs$x, res)
```

We also compared the omnibus estimates to independent estimates

```{r rq2-figure-2}
p1 %+%
  drop_na(tmp) +
  aes(col = Model) +
  facet_wrap("Game")
```

## Nonlinear models

We also investigated potential nonlinear relations between game time and wellbeing. We did so by fitting a model with and without a smooth term, and using AIC to compare the models.

```{r nonlinear-figure}
dat %>%
  select(Game, Spane_balance, Hours10, Active_play10) %>% 
  pivot_longer(contains("10")) %>% 
  ggplot(aes(value, Spane_balance)) +
  geom_point() +
  geom_smooth(method = "lm", aes(col = "linear")) +
  geom_smooth(method = "gam", aes(col = "GAM")) +
  facet_grid(name~Game, scales = "free")
```

```{r}
library(mgcv)
dat %>% 
  select(Game, Spane_balance, Hours10, Active_play10) %>% 
  pivot_longer(contains("10"), names_to = "Variable") %>%
  group_by(Game, Variable) %>% 
  nest() %>% 
  mutate(linear = map(data, ~gam(Spane_balance ~ value, data = .x))) %>% 
  mutate(smooth = map(data, ~gam(Spane_balance ~ s(value), data = .x))) %>% 
  pivot_longer(linear:smooth, names_to = "Model") %>% 
  mutate(AIC = map_dbl(value, AIC)) %>% 
  select(Game, Variable, Model, AIC) %>% 
  pivot_wider(names_from = Model, values_from = AIC) %>% 
  mutate(Difference = linear-smooth) %>% 
  kable(digits = 1)
```

## System information

```{r}
sessionInfo()
```
